Move chats to interface
