+++ GPTifier/src/main.cpp
@@ -1,8 +1,8 @@
-#include "command_chats.hpp"
 #include "command_fine_tune.hpp"
 #include "command_test.hpp"
 #include "configs.hpp"
 #include "help_messages.hpp"
+#include "interface/command_chats.hpp"
 #include "interface/command_costs.hpp"
 #include "interface/command_embed.hpp"
 #include "interface/command_files.hpp"
